Implement CRUD operations for Disease entity in DiseaseImpl; update DiseaseService interface to use Long for IDs

+++ src/main/java/com/schoolhealth/schoolmedical/service/DiseaseService.java
@@ -11,8 +11,8 @@ import java.util.Optional;
 public interface DiseaseService {
     DiseaseResponse createDisease(DiseaseRequest request);
     DiseaseResponse updateDisease(Long id, DiseaseRequest request);
-    void deleteDisease(int id);
-    DiseaseResponse getDiseaseById(int id);
+    void deleteDisease(Long id);
+    DiseaseResponse getDiseaseById(Long id);
 //    List<DiseaseResponse> getAllDiseases();
     Page<DiseaseResponse> getAllDiseases(int pageNo, int pageSize, Boolean isActive);
